refactor: refactoring the folders
