+++ src/components/Sidebar/styles.ts
@@ -3,4 +3,4 @@ export const textStyle = ' font-bold text-base px-9';
 export const textStyle2 = ' font-normal text-sm px-9';
 export const buttonStyle = 'flex justify-start text-greenText font-semibold text-base px-7 w-full';
 export const belowStyle = 'flex flex-row justify-center items-center bg-grayBg h-90 bottom-0 left-0 mt-auto';
-export const buttonStyleSidebar = 'items-center hover:opacity-75 transition-all duration-300';
+export const divGeneral = 'flex h-screen w-80 bg-Input2 flex-col'
