+++ src/pages/configuracao/styles.ts
@@ -4,3 +4,4 @@ export const textStyle2 = ' font-normal text-sm px-9';
 export const textStyle3 = 'font-semibold text-2xl pt-9';
 export const buttonStyle = 'flex justify-start text-greenText font-semibold text-base px-7 w-full';
 export const belowStyle = 'flex flex-row justify-center items-center bg-grayBg h-90 bottom-0 left-0 mt-auto';
+export const divGeneral ='flex bg-white h-screen w-full items-center flex-col px-28 py-20';
