+++ src/pages/login/index.tsx
@@ -3,7 +3,7 @@ import React, { useState } from 'react';
 import { Button } from '../../components/Button';
 import { Input } from '../../components/Input';
 import { Stack } from '../../components/Stack';
-import { selectStyle } from './styles';
+import { divGeneral, selectStyle, textTitle } from './styles';
 
 export default function Login() {    
     const [buttonStartup, setButtonStartup] = useState(true);
@@ -44,12 +44,11 @@ export default function Login() {
         <Stack bg='bg-white'>
             <div className="h-screen w-3/5 bg-agro bg-cover bg-center"/>
 
-            <div className="flex justify-center items-center bg-white h-screen w-full grid direction-column">
-                <div>
-                
+            <div className={divGeneral}>
+                <div>                
                     <div className="w-36 h-8 bg-no-repeat bg-agroLogo "/>
 
-                    <h1 className="font-semibold text-4xl mt-3">Login</h1>
+                    <h1 className={textTitle}>Login</h1>
 
                     <div className={selectStyle}>                    
                         <Button
