+++ src/pages/login/styles.ts
@@ -1,1 +1,3 @@
 export const selectStyle = 'flex bg-optionWhite h-12 w-471 rounded-full direction-row items-center justify-between px-[.5rem] mt-5';
+export const divGeneral = 'flex justify-center items-center bg-white h-screen w-full grid direction-column';
+export const textTitle = 'font-semibold text-4xl mt-3';
