+++ src/pages/minhaContaAdmin/styles.ts
@@ -1,6 +1,3 @@
-export const textTitle = 'text-textBlack13 font-bold text-base px-9';
-export const textStyle = ' font-bold text-base px-9';
-export const textStyle2 = ' font-normal text-sm px-9';
-export const textStyle3 = 'font-semibold text-2xl';
-export const buttonStyle = 'flex justify-start text-greenText font-semibold text-base px-7 w-full';
-export const belowStyle = 'flex flex-row justify-center items-center bg-grayBg h-90 bottom-0 left-0 mt-auto';
+export const textTitle = 'font-semibold text-4xl';
+export const textStyle3 = 'font-semibold text-2xl pt-9';
+export const divGeneral = 'flex justify-center items-center bg-white h-screen w-full grid direction-column pt-8';
